requirements.txt
Django>=5.1.0,<6.0.0
djangorestframework>=3.15.0,<4.0.0
python-decouple>=3.8
django-cors-headers>=4.3.0